fix(15.A): object field cells must be intptr_t, not long (Windows LLP64)
CI caught test_teko_object_holds_handle failing on Windows MSVC (x86_64 + arm64,
native + the windows-stress sanitizer job): a field storing another object's
handle came back sign-extended (0xFFFFFFFF...) because `long` is only 32-bit on
Windows LLP64 while pointers are 64-bit — the cell truncated the handle.

Widen the field cells + teko_object_set/get to `intptr_t` (pointer width on
every ABI: LP64, LLP64, and wasm32), so a field genuinely holds either an int
or another object's handle everywhere. The teko_rt_object_* wrappers keep the
register-width `long` OP_OBJ ABI (correct for the LP64 native runner targets and
wasm32; the Windows PE-COFF native runner remains future work). Unix/wasm output
unchanged. Suite 213/213; ASan/UBSan (both paths) + TSan clean; native (7,70) and
WASM ([7,70]) proofs intact.

+++ src/runtime/teko_object.c
@@ -6,8 +6,8 @@
 // for both the native runner (teko_rt_object_* wrappers) and the wasm32 reactor.
 
 struct TekoObject {
-    int  nfields;
-    long cells[]; // flexible array: nfields zero-initialized register-width cells
+    int      nfields;
+    intptr_t cells[]; // flexible array: nfields zero-initialized POINTER-width cells (Windows-safe)
 };
 
 TekoObject* teko_object_new(int nfields) {
@@ -15,7 +15,7 @@ TekoObject* teko_object_new(int nfields) {
     if (nfields > TEKO_OBJECT_MAX_FIELDS) nfields = TEKO_OBJECT_MAX_FIELDS;
     // calloc zero-initializes the header + every cell (the Windows wild-free lesson: never leave
     // an allocation's fields unset).
-    TekoObject* o = (TekoObject*)calloc(1, sizeof(TekoObject) + (size_t)nfields * sizeof(long));
+    TekoObject* o = (TekoObject*)calloc(1, sizeof(TekoObject) + (size_t)nfields * sizeof(intptr_t));
     if (!o) return NULL;
     o->nfields = nfields;
     return o;
@@ -25,12 +25,12 @@ void teko_object_free(TekoObject* o) {
     free(o);
 }
 
-void teko_object_set(TekoObject* o, int idx, long value) {
+void teko_object_set(TekoObject* o, int idx, intptr_t value) {
     if (!o || idx < 0 || idx >= o->nfields) return; // defensive: never write OOB
     o->cells[idx] = value;
 }
 
-long teko_object_get(const TekoObject* o, int idx) {
+intptr_t teko_object_get(const TekoObject* o, int idx) {
     if (!o || idx < 0 || idx >= o->nfields) return 0; // defensive: 0 for OOB/NULL
     return o->cells[idx];
 }
